顶部栏header功能oo化
1、包括保存、全屏、modelView等功能
