+++ views/workspace.ejs
@@ -23,7 +23,7 @@
     </button>
   </div>
   <!-- 顶部项目名称栏 -->
-  <div class="row stigmod-hide-when-full-screen">
+  <div class="row stigmod-hide-when-full-screen" id="stigmod-header">
     <div class="col-xs-8">
       <h1 id="stigmod-model-title">
         <span class="fa fa-cube" id="stigmod-model-title-icon"></span>
